requirements.txt
Django==2.2.4
django-debug-toolbar==2.0
django-filter==2.2.0
djangorestframework==3.10.2
factory-boy==2.12.0
Faker==2.0.1
Markdown==3.1.1
Pillow==6.1.0
psycopg2-binary==2.8.3
python-dateutil==2.8.0
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
